# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.Card:
    properties:
      cost:
        example: 2
        type: integer
      description:
        example: Deal 8 damage and apply Vulnerable
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Code Slash
        type: string
      type:
        enum:
        - action
        - event
        - power
        example: action
        type: string
    type: object
...
  main.ErrorResponse:
    properties:
      error:
        example: Not found
        type: string
      message:
        example: The requested resource was not found
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pixelgame.io
    name: Pixel Game Backend Team
    url: https://github.com/HariFatherKR/pixel-game-backend/issues
  description: Vibe 코딩 기반 사이버펑크 덱 빌딩 카드 게임의 백엔드 API 서버입니다. 프론트엔드 코드 로직과 게임플레이를 연결하여 카드 사용 시 실제 코드가 실행되는 혁신적인 게임입니다.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/HariFatherKR/pixel-game-backend
  title: Pixel Game - 사이버펑크 덱 빌딩 카드 게임 API
  version: "0.1.0"
paths:
...
  /cards/{id}:
    get:
      consumes:
      - application/json
      description: Get details of a specific card
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Card'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get card by ID
      tags:
      - cards
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"